requirements.txt
python-dateutil
werkzeug
autocorrect==0.4.4
dataclasses
dataclasses_json
editdistance==0.5.3
flask
Flask
imutils==0.5.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pytesseract==0.3.0
pyocr
pytest
scikit-image
setuptools
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
pyyaml